Ultimo Commit

+++ src/main/java/co/edu/ucentral/tarjetac/dto/SolicitudesDto.java
@@ -16,12 +16,12 @@ public class SolicitudesDto implements Serializable {
     private String nombre;
     @NotBlank(message = "El apellido es obligatorio")
     private String apellido;
-    private int celular;
+    private long celular;
     @NotBlank(message = "El correo es obligatorio")
     private String correo;
-    private int cedula;
-    private int salario;
-    private int gastos;
+    private long cedula;
+    private long salario;
+    private long gastos;
     @NotBlank(message = "El comentario es obligatorio")
     private String comentario;
 }
